subscribe channel check

+++ bot/bot.py
@@ -5,6 +5,7 @@ import telebot
 
 BOT_TOKEN = config.BOT_TOKEN
 GROUP_ID = config.GROUP_ID
+TARGET_CHANNEL = config.TARGET_CHANNEL
 bot = telebot.TeleBot(BOT_TOKEN)
 
 
@@ -12,6 +13,7 @@ bot = telebot.TeleBot(BOT_TOKEN)
 @bot.message_handler(content_types = ['text'])
 def spam_security(message):
     
+    message_is_deleted = False
     if str(message.chat.id) in GROUP_ID\
         and str(message.from_user.username) not in config.SUPERUSERS:
     
@@ -20,16 +22,32 @@ def spam_security(message):
             bot.delete_message(message.chat.id, message.message_id)
             utils.send_info_message_and_delete_old_message(
                 message=message, 
-                info=utils.ErrorMessage.phone,
+                info=utils.ErrorMessage.phone(),
                 )
+            utils.log(message=message)
+            message_is_deleted = True
         
         if str(message.chat.id) == GROUP_ID[1]:    
-            if not utils.Check.max_lenght(message):
+            if not utils.Check.max_lenght(message) and message_is_deleted:
                 bot.delete_message(message.chat.id, message.message_id)
                 utils.send_info_message_and_delete_old_message(
                     message=message,
                     info=utils.ErrorMessage.lenght(message)
                 )
+                message_is_deleted = True
+
+            try:
+                response = bot.get_chat_member(chat_id=TARGET_CHANNEL, user_id=int(message.from_user.id))
+                if response.status != ('member' or "creator") and message_is_deleted != False:
+                    bot.delete_message(message.chat.id, message.message_id)
+                    utils.send_info_message_and_delete_old_message(
+                        message=message,
+                        info=utils.ErrorMessage.subscribe_channel()
+                    )
+                    message_is_deleted = True
+
+            except:
+                print('user is not member')
